+++ Kernel_23/include/CGAL/Kernel/global_functions_internal_2.h
@@ -24,10 +24,11 @@
 
 #include <CGAL/basic.h>
 #include <CGAL/Dimension.h>
-#include <boost/utility/enable_if.hpp>
 #include "boost/mpl/equal_to.hpp"
 #include <boost/mpl/int.hpp>
 #include <boost/mpl/integral_c.hpp>
+
+#include <type_traits>
 
 namespace CGAL {
 
@@ -273,12 +274,11 @@ compare_angle_with_x_axis(const typename K::Direction_2& d1,
 
 template <class K, class T1, class T2, class T3>
 inline
-typename boost::enable_if<
+std::enable_if_t<
   boost::mpl::equal_to<boost::mpl::integral_c<int,
                                               Ambient_dimension<T1>::type::value>,
-                       boost::mpl::integral_c<int, 2> >,
+                       boost::mpl::integral_c<int, 2> >::value,
   typename K::Comparison_result>
-::type
 compare_distance(const T1 &o1,
                  const T2 &o2,
                  const T3 &o3, const K& k)
@@ -288,12 +288,11 @@ compare_distance(const T1 &o1,
 
 template <class K, class T1, class T2, class T3, class T4>
 inline
-typename boost::enable_if<
+std::enable_if_t<
   boost::mpl::equal_to<boost::mpl::integral_c<int,
                                               Ambient_dimension<T1>::type::value>,
-                       boost::mpl::integral_c<int, 2> >,
+                       boost::mpl::integral_c<int, 2> >::value,
   typename K::Comparison_result>
-::type
 compare_distance(const T1 &o1,
                  const T2 &o2,
                  const T3 &o3,