+++ Nef_2/include/CGAL/Nef_2/PM_overlayer.h
@@ -32,7 +32,7 @@
 #include <CGAL/assertions.h>
 #include <CGAL/use.h>
 
-#include <boost/type_traits/is_same.hpp>
+#include <type_traits>
 
 #ifndef CGAL_USE_LEDA
 #define LEDA_MEMORY(t)
@@ -959,7 +959,7 @@ bool is_forward_edge(const Const_decorator& N,
 
 void assert_type_precondition() const
 { typename PM_decorator_::Point p1; Point p2;
-  CGAL_static_assertion((boost::is_same<typename PM_decorator_::Point, Point>::value)); }
+  CGAL_static_assertion((std::is_same<typename PM_decorator_::Point, Point>::value)); }