+++ NewKernel_d/include/CGAL/NewKernel_d/Vector/array.h
@@ -12,7 +12,6 @@
 #ifndef CGAL_VECTOR_ARRAY_H
 #define CGAL_VECTOR_ARRAY_H
 #include <boost/type_traits/is_arithmetic.hpp>
-#include <boost/utility/enable_if.hpp>
 #include <CGAL/Dimension.h>
 #include <CGAL/NewKernel_d/utils.h>
 #include <CGAL/array.h>