+++ Number_types/include/CGAL/mpq_class.h
@@ -36,7 +36,7 @@
 
 #define CGAL_CHECK_GMP_EXPR_MPQ_CLASS                                     \
     CGAL_static_assertion(                                                \
-            (::boost::is_same< ::__gmp_expr< T , T >,Type>::value ));
+            (::std::is_same< ::__gmp_expr< T , T >,Type>::value ));
 
 namespace CGAL {